+++ SetUpClass_Cata.lua
@@ -4,6 +4,8 @@ local addonName, Engraved = ...
 
 local RuneFrame = EngravedRuneFrame
 
+EngravedFrame:RegisterUnitEvent("PLAYER_SPECIALIZATION_CHANGED", "player")
+
 function Engraved:SetupClass()
 	RuneFrame:RegisterEvent("PLAYER_REGEN_DISABLED")
 	RuneFrame:RegisterEvent("PLAYER_REGEN_ENABLED")
@@ -23,10 +25,8 @@ function Engraved:SetupClass()
 		Engraved.Druid:SetupClassic()
 	elseif class == "PALADIN" then
 		Engraved.Paladin:Setup()
-	--[[
 	elseif class == "WARLOCK" then
 		Engraved.Warlock:Setup()
-	]]--
 	else
 		RuneFrame.inUse = false
 		RuneFrame:UnregisterAllEvents()