+++ SetUpClass_Retail.lua
@@ -7,7 +7,6 @@ local RuneFrame = EngravedRuneFrame
 
 local GetSpecialization = GetSpecialization or GetPrimaryTalentTree
 
--- Events not in Classic
 EngravedFrame:RegisterUnitEvent("PLAYER_SPECIALIZATION_CHANGED", "player")
 
 function Engraved:SetupClass()