+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### v2.3.0 (2024-XX-XX)
 * Added: Now supports Frost Mage icicles (Retail). Thanks, huiyuanai709! 
+* Added: Now supports Paladin holy power in Cataclysm classic
 * Fixed: All Paladin specializations can now use holy power
 * Updated Table of Contents for patch 11.0.2